refactor: refactor used pinia store method

feat: refactor
